Add $manage_backends parameter.
This adds a $manage_backends parameter that is used to determine if the
various supported statsd backend are managed by this class.  The default
value is 'true', and thus the default behavior is unchanged.  Setting
the value to 'false' will disable calling statsd::backends, and force
admins to manage the backends via other means.